-Clean and Categorize code (the tabular part)
