+++ src/Covid19/DaysFromAParticularDate.java
@@ -4,23 +4,23 @@ import java.io.*;
 import java.util.*;
 import java.text.ParseException;
 
-import static Covid19.UserInput.userInputRequest;
+import static Covid19.UserInitialInput.userInputRequest;
 import static Covid19.ReadWriteCsvData.readCsvRow;
 import static Covid19.Metrics.*;
 
 class DaysFromAParticularDate {
     public static void Process() throws ParseException, IOException {
-        UserInput userDataInput = userInputRequest();
-
+        UserInitialInput userDataInput = userInputRequest();
         String pathToNewCSV = userDataInput.pathToNewCsv;
         String location = userDataInput.location;
         String chosenDate = userDataInput.chosenDate;
         int dayAway = userDataInput.dayAway;
         int option = userDataInput.option;
+        //filter and get the required Data
+        ArrayList<CovidData> bigGroup = readCsvRow(pathToNewCSV, location, chosenDate, dayAway);
 
         if (option == 1) {
-            //read Data and get Metric
-            ArrayList<CovidData> bigGroup = readCsvRow(pathToNewCSV, location, chosenDate, dayAway);
+            //get metric option
             int metricOption = metricUserInput();
             //Print Data for users
             System.out.println("\n--- Data List ---");
@@ -30,19 +30,11 @@ class DaysFromAParticularDate {
             //metricsArr for saving the metric result
             ArrayList<Long> metricsArr = new ArrayList<Long>();
             metricDisplay(metricOption, bigGroup, metricsArr);
-            //Choose a display
-            System.out.println("\nChoose one way to display");
-            System.out.println("1. Tabular display ");
-            System.out.println("2. Chart display ");
-            int DisplayOption;
-
-//            do {
-//                Scanner input = new Scanner(System.in);
-//                System.out.print("Please enter the way you want: ");
-//                DisplayOption = input.nextInt();
-//            } while (DisplayOption != 1 && DisplayOption != 2 );
-//            if (DisplayOption == 1)
-//                new TabularDisplay(chosenDate,endDate,metricsArr);
+            //Display Option and Table
+            int DisplayOption = UserDisplayInput.optionDisplayInput();
+            String endDate = TimeRelatedFunctions.displayStartEndDate(chosenDate, dayAway);
+            if (DisplayOption == 1)
+                new TabularDisplay(chosenDate,endDate,metricsArr);
 
         } else if (option == 2) {
             int groups;
@@ -53,19 +45,10 @@ class DaysFromAParticularDate {
             } while (groups > dayAway);
             //GETTING THE NUMBER OF NEEDED-SPLITTED GROUPS
             ArrayList<Integer> groupsSplittedArr = DayGroupSplitting.splitGroupsEqually(dayAway, groups);
-            //read Data and get Metric
-            ArrayList<CovidData> bigGroup = readCsvRow(pathToNewCSV, location, chosenDate, dayAway);
+            //get Metric option
             int metricOption = metricUserInput();
-//            //Choose a display
-            System.out.println("\nChoose one way to display");
-            System.out.println("1. Tabular display ");
-            System.out.println("2. Chart display ");
-            int DisplayOption;
-            do {
-                Scanner input = new Scanner(System.in);
-                System.out.print("Please enter the way you want: ");
-                DisplayOption = input.nextInt();
-            } while (DisplayOption != 1 && DisplayOption != 2 );
+            //Display Option and Table
+            int DisplayOption = UserDisplayInput.optionDisplayInput();
             //Put data into groups and print out
             DayGroupSplitting.putDataInGroup(bigGroup, groupsSplittedArr, chosenDate, metricOption, DisplayOption);
 
@@ -78,20 +61,10 @@ class DaysFromAParticularDate {
             } while (daysPerGroup > dayAway && daysPerGroup > 1);
             //GETTING THE NUMBER OF NEEDED-SPLITTED GROUPS
             ArrayList<Integer> groupsSplittedArr = DayGroupSplitting.splitEqualDays(dayAway, daysPerGroup);
-            //read Data and get Metric
-            ArrayList<CovidData> bigGroup = readCsvRow(pathToNewCSV, location, chosenDate, dayAway);
+            //get Metric option
             int metricOption = metricUserInput();
-            //Choose way of displaying
-            System.out.println("\nChoose one way to display");
-            System.out.println("1. Tabular display ");
-            System.out.println("2. Chart display ");
-            int DisplayOption;
-            do {
-                Scanner input = new Scanner(System.in);
-                System.out.print("Please enter the way you want: ");
-                DisplayOption = input.nextInt();
-            } while (DisplayOption != 1 && DisplayOption != 2 );
-
+            //Display Option and Table
+            int DisplayOption = UserDisplayInput.optionDisplayInput();
             //Put data into groups and print out
             DayGroupSplitting.putDataInGroup(bigGroup, groupsSplittedArr, chosenDate, metricOption, DisplayOption);
