+++ src/Covid19/FilteringSplitingFunc.java
@@ -1,12 +1,9 @@
 package Covid19;
 
-import java.text.DateFormat;
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.*;
 
 import static Covid19.Metrics.metricDisplay;
-
 
 class DateLocationFiltering {
     public static CovidData getDataFromDate(String chosenDate, String endDate, CovidData dataRow) throws ParseException {
@@ -43,64 +40,6 @@ class DateLocationFiltering {
             dataRow = null;
         }
         return dataRow;
-    }
-}
-
-class TimeRelatedFunctions {
-    public static String displayStartEndDate(String chosenDate, int dayAway) {
-        //Specifying date format
-        SimpleDateFormat sdf = new SimpleDateFormat("M/d/yyyy");
-        Calendar c = new GregorianCalendar();
-        try {
-            Date date = sdf.parse(chosenDate);
-            //Setting the date to the given date
-            c.setTime(date);
-//            System.out.println(date);
-        } catch (ParseException e) {
-            System.out.println("invalid input");
-            e.printStackTrace();
-        }
-
-        //Number of Days to add
-        c.add(Calendar.DAY_OF_MONTH, dayAway);
-        //Day after adding the days to the input date
-        String endDate = sdf.format(c.getTime());
-        //Displaying the new Date after addition of Days
-        System.out.println("Date after Addition: " + endDate);
-        return endDate;
-    }
-
-    public static ArrayList<Date> getDatesBetween(String startDate, String endDate) throws ParseException {
-        //convert String to Date
-        SimpleDateFormat sdf = new SimpleDateFormat("M/d/yyyy");
-        Date sDate = sdf.parse(startDate);
-        Date eDate = sdf.parse(endDate);
-
-        ArrayList<Date> datesInRange = new ArrayList<>();
-        Calendar c = new GregorianCalendar();
-        c.setTime(sDate);
-
-        Calendar endCalendar = new GregorianCalendar();
-        endCalendar.setTime(eDate);
-
-        while (c.before((endCalendar))) {
-            Date result = c.getTime();
-            datesInRange.add(result);
-            c.add(Calendar.DATE, 1);
-        }
-        return datesInRange;
-    }
-
-    public static ArrayList<String> convertDateToString(ArrayList<Date> ArrList) {
-        DateFormat df = new SimpleDateFormat("M/d/yyyy");
-        ArrayList<String> dateToStrArr = new ArrayList<String>();
-
-        for (int i = 0; i < ArrList.size(); i++) {
-            Date DateInArr = ArrList.get(i);
-            String dateToString = df.format(DateInArr);
-            dateToStrArr.add(dateToString);
-        }
-        return dateToStrArr;
     }
 }
 
@@ -225,6 +164,7 @@ class DayGroupSplitting {
         return groupsSplittedArr;
     }
 }
+
